+++ modules/k8s_eks_addons/variables.tf
@@ -40,4 +40,10 @@ variable "aws_load_balancer_controller_config" {
     helm_version    = string
     chart_values    = string
   })
+variable "s3_csi_config" {
+  type = object({
+    enable               = optional(bool, false)
+    configuration_values = optional(string, null)
+  })
+  description = "Input configuration for AWS EKS add-on aws-mountpoint-s3-csi-driver."
 }